Extract wheel paging into wheel-pager module
Move the velocity-driven wheel paging logic out of overlay.ts into a new src/ui/single-page/wheel-pager.ts. Introduces createWheelPager with a small config API (getPswp, getImageCount, isPageLoading, onEdgeForward, onEdgeBackward) and exposes onWheel/stop. overlay.ts now imports and delegates wheel events and destroy handling to the pager. Existing behavior (decay, velocity thresholds, loading-speed bump, and edge callbacks that load next/prev pages) is preserved.
